requirements.txt

fastapi
uvicorn
pydantic
numpy
pandas
gspread
google-auth
xgboost
yfinance
